@@ -2,22 +2,23 @@
   "name": "en-words",
   "version": "0.0.2",
   "devDependencies": {
-    "react-scripts": "0.9.5"
+    "react-scripts": "1.0.17"
   },
   "dependencies": {
-    "antd": "^2.6.4",
-    "axios": "^0.15.3",
+    "antd": "^3.0.3",
+    "axios": "^0.17.1",
     "bootstrap": "^3.3.7",
-    "firebase": "^3.7.4",
-    "react": "^15.5.4",
-    "react-bootstrap": "^0.30.7",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.2",
-    "react-router": "^3.0.0",
-    "redux": "^3.6.0",
-    "redux-form": "^6.6.2",
-    "redux-logger": "^2.7.4",
-    "redux-promise-middleware": "^4.2.0",
+    "firebase": "^4.8.1",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-bootstrap": "^0.31.5",
+    "react-dom": "^16.2.0",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
+    "redux": "^3.7.2",
+    "redux-form": "^7.2.0",
+    "redux-logger": "^3.0.6",
+    "redux-promise-middleware": "^5.0.0",
     "redux-thunk": "^2.1.0"
   },
   "scripts": {
