@@ -1,26 +1,25 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Button } from 'antd';
 import './GroupToolBar.css';
-
-const ButtonGroup = Button.Group;
 
 const GroupToolBar = (props) => {
     const { onNewGroup, onEditGroup, onDeleteGroup, selectedGroup } = props;
 
     return (
-        <ButtonGroup id="groupButtons" className="align-right padding-top-5px">
+        <Button.Group id="groupButtons" className="align-right padding-top-5px">
             <Button onClick={onNewGroup} icon="plus-circle-o" />
             <Button onClick={onEditGroup} disabled={!selectedGroup} icon="edit" />
             <Button onClick={onDeleteGroup} disabled={!selectedGroup} icon="delete" />
-        </ButtonGroup>
+        </Button.Group>
     )
 };
 
 GroupToolBar.propTypes = {
-    onNewGroup: React.PropTypes.func.isRequired,
-    onEditGroup: React.PropTypes.func.isRequired,
-    onDeleteGroup: React.PropTypes.func.isRequired,
-    selectedGroup: React.PropTypes.object
+    onNewGroup: PropTypes.func.isRequired,
+    onEditGroup: PropTypes.func.isRequired,
+    onDeleteGroup: PropTypes.func.isRequired,
+    selectedGroup: PropTypes.object
 };
 
 export default GroupToolBar;
