@@ -1,9 +1,12 @@
 import React, { Component } from 'react';
+import { Switch, Route } from 'react-router-dom'
 
 import Header from './../containers/HeaderContainer';
 import Groups from "../../groups/containers/GroupPageContainer";
 
 import './App.css';
+import Dashboard from "../../dashboard/components/Dashboard";
+import Words from "../../words/containers/WordPageContainer";
 
 class App extends Component {
 
@@ -17,7 +20,11 @@ class App extends Component {
                         <Groups />
                     </div>
                     <div className="col-sm-9">
-                        {this.props.children}
+                        <Switch>
+                            <Route path='/dashboard' component={Dashboard} />
+                            <Route path='/words/:groupId' component={Words} />
+                            <Route exac path='/' component={Dashboard} />
+                        </Switch>
                     </div>
                 </div>
             </div>
