@@ -1,25 +1,19 @@
-import React, { PropTypes } from 'react';
-import { Link } from 'react-router';
+import React from 'react';
+import { Link } from 'react-router-dom';
 
 import { PageHeader } from 'react-bootstrap'
 
 
 import { APP_VERSION } from '../../common/AppSettings'
 
-const Header = (props) => {
-
-    const { onHeaderClick } = props;
+const Header = () => {
 
     return (
         <PageHeader>
-            <Link onClick={onHeaderClick}>Learn-Words</Link> <small>(v. {APP_VERSION})</small>
+            <Link to="/">Learn-Words</Link> <small>(v. {APP_VERSION})</small>
         </PageHeader>
     );
 
-};
-
-Header.propTypes = {
-    onHeaderClick: PropTypes.func.isRequired
 };
 
 export default Header;
